login successfully returns a token in the post request
