+++ backend/src/main/java/cps/models/requestBodies/LoginRequestBody.java
@@ -1,0 +1,14 @@
+package cps.models.requestBodies;
+
+public class LoginRequestBody {
+  private String email;
+  private String password;
+
+  public String getEmail() {
+    return email;
+  }
+
+  public String getPassword() {
+    return password;
+  }
+}
